[REFACTOR] Update Network and Server request url & divide components and style
네트워크 및 서버 요청 url 수정 및 컴포넌트와 스타일 변경
