--- client/src/components/mypage/MyInfo.jsx
@@ -1,123 +1,0 @@
-import React, { useState } from 'react';
-import styled from 'styled-components';
-import { useDispatch, useSelector } from 'react-redux';
-import NameInput from './Myinfo/NameInput';
-import IdInput from './Myinfo/IdInput';
-import EmailInput from './Myinfo/EmailInput';
-import ProfileImg from './MyProfile/ProfileImg';
-import { UpdateBtn } from '../../styles/common/buttons';
-import { updateImage, updateUser } from '../../redux/login/action';
-import ImgInput from './Myinfo/ImgInput';
-
-const Wrapper = styled.form`
-  width: 70%;
-  height: 90%;
-  display: flex;
-  margin-top: 50px;
-`;
-
-const ImgForm = styled.div`
-  width: 50%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 40px;
-  &:first-child {
-    padding-top: 30px;
-  }
-  box-sizing: border-box;
-`;
-
-const InfoForm = styled.div`
-  width: 50%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 25px;
-`;
-
-function MyInfo () {
-  const user = useSelector((state) => state.loginReducer);
-  const dispatch = useDispatch();
-  const [userInfo, setUserInfo] = useState({
-    id: user.id,
-    nickname: user.name,
-    email: user.email,
-    imgFiles: "",
-    encodeImg: "",
-  });
-  
-  const [myMessage, setMyMessage] = useState({
-    id: "",
-    nickname: "",
-  });
-
-  const [success, setSuccess] = useState({
-    id: false,
-    nickname: false,
-  })
-
-  const handleInputChange = (evt) => {
-    const name = evt.target.name;
-    setUserInfo({ ...userInfo, [name]: evt.target.value});
-  };
-
-  const handleInfoUpdate = (evt) => {
-    evt.preventDefault();
-
-    const formdata = new FormData();
-    formdata.append('email', userInfo.email);
-    if (userInfo.imgFiles !== "") formdata.append('file', userInfo.imgFiles[0]);
-
-    const lastIdCheck = userInfo.id.indexOf("\b");
-    const lastNameCheck = userInfo.nickname.indexOf("\b");
-    if (!(user.id === userInfo.id || user.name === userInfo.nickname) && (lastIdCheck !== 0 && lastNameCheck !== 0)) {
-      formdata.append('userId', userInfo.id);
-      formdata.append('nickname', userInfo.nickname);
-      dispatch(updateUser(formdata));
-    } else {
-      dispatch(updateImage(formdata));
-    }
-  }
-
-  return (
-    <Wrapper method='POST' encType='multipart/form-data'>
-      <ImgForm>
-        <ProfileImg 
-          userInfo={userInfo}
-        />
-        <ImgInput 
-          userInfo={userInfo}
-          setUserInfo={setUserInfo}
-        />
-      </ImgForm>
-      <InfoForm>
-        <IdInput
-          userId={userInfo.id}
-          onInputChange={handleInputChange}
-          myMessage={myMessage}
-          setMyMessage={setMyMessage}
-          success={success}
-          setSuccess={setSuccess}
-        />
-        <NameInput
-          userName={userInfo.nickname}
-          onInputChange={handleInputChange}
-          myMessage={myMessage}
-          setMyMessage={setMyMessage}
-          success={success}
-          setSuccess={setSuccess}
-        />
-        <EmailInput 
-          userEmail={userInfo.email}
-        />
-        <UpdateBtn type="submit" onClick={handleInfoUpdate}> 수정 </UpdateBtn>
-      </InfoForm>
-    </Wrapper>
-  )
-}
-
-export default MyInfo;
