+++ client/src/components/mypage/MyProfile/MyProfile.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import MyProfileStyle from './MyProfile.style';
+import { useSelector } from 'react-redux';
+
+function MyProfile () {
+  const { social } = useSelector((state) => state.loginReducer);
+
+  return (
+    <MyProfileStyle 
+      social={social}
+    />
+  )
+}
+
+export default MyProfile;
