+++ client/src/components/mypage/Myinfo/InfoInput/EmailInput.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { MyInfoInput } from '../../../styles/common/input';
-import { InfoBox, InfoLabel } from '../../../styles/mypage/myinfo';
+import { MyInfoInput } from '../../../../styles/common/input';
+import { InfoBox, InfoLabel } from '../../../../styles/mypage/myinfo';
 
 function EmailInput ({userEmail}) {
   return (
