+++ client/src/components/mypage/PasswordChange/PasswordChange.jsx
@@ -1,29 +1,10 @@
 import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
-import styled from 'styled-components';
 import { putChangePassWord } from '../../../network/mypage/http';
-import { ChangePwBtn, CancelBtn } from '../../../styles/common/buttons';
 import { pwValidation } from '../../../utils/validation';
-import PwInput from '../Myinfo/PwInput';
+import PasswordChangeStyle from './PasswordChange.style';
 
-const Wrapper = styled.form`
-  width: 50vw;
-  height: 60vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 20px;
-`;
-
-const Title = styled.h2`
-  width: 70%;
-  height: 70px;
-  line-height: 70px;
-  font-size: 28px;
-  font-weight: bold;
-`;
-
-function PwChange () {
+function PasswordChange () {
   const user = useSelector((state) => state.loginReducer);
   const [password, setPassword] = useState({
     oldPassword: "",
@@ -53,15 +34,11 @@ function PwChange () {
   }
 
   return (
-    <Wrapper>
-      <Title> 비밀번호 변경 </Title>
-      <PwInput
-        onInputChange={handleInputChange}
-      />
-      <ChangePwBtn onClick={handlePwSubmit}> 확인 </ChangePwBtn>
-      <CancelBtn> 취소 </CancelBtn>
-    </Wrapper>
+    <PasswordChangeStyle
+      onInputChange={handleInputChange}
+      onPwSubmit={handlePwSubmit}
+    />
   );
 }
 
-export default PwChange;
+export default PasswordChange;
